+++ src/main/java/com/github/jlangch/venice/impl/docgen/cheatsheet/DocGenerator.java
@@ -2034,6 +2034,8 @@ public class DocGenerator {
 		final DocSection util = new DocSection("Util", id());
 		all.addSection(util);
 		util.addItem(getDocItem("trace/trace-str-limit"));
+		util.addItem(getDocItem("trace/tee->"));
+		util.addItem(getDocItem("trace/tee->>"));
 
 		return section;
 	}